tách api
