+++ src/api/index.js
@@ -1,0 +1,4 @@
+// Export all API modules
+export { authAPI } from './auth';
+export { servicesAPI } from './services';
+export { userAPI } from './user';
