+++ src/api/user.js
@@ -1,0 +1,9 @@
+import { apiFetch } from '../utils/api';
+
+// User API functions
+export const userAPI = {
+  // Lấy thông tin user hiện tại
+  getMyInfo: async (token) => {
+    return await apiFetch('/v1/users/myInfo', { token });
+  }
+};
